refactor(AWS Deploy): Introduce `getExecuteChangeSetParams` util
