+++ lib/plugins/aws/deploy/index.js
@@ -8,6 +8,7 @@ const checkForChanges = require('./lib/checkForChanges');
 const monitorStack = require('../lib/monitorStack');
 const checkIfBucketExists = require('../lib/check-if-bucket-exists');
 const getCreateChangeSetParams = require('../lib/get-create-change-set-params');
+const getExecuteChangeSetParams = require('../lib/get-execute-change-set-params');
 const waitForChangeSetCreation = require('../lib/wait-for-change-set-creation');
 const createStack = require('./lib/createStack');
 const cleanupS3Bucket = require('./lib/cleanupS3Bucket');
@@ -46,7 +47,8 @@ class AwsDeploy {
       monitorStack,
       checkIfBucketExists,
       waitForChangeSetCreation,
-      getCreateChangeSetParams
+      getCreateChangeSetParams,
+      getExecuteChangeSetParams
     );
 
     this.getFileStats = memoize(this.getFileStats.bind(this), { promise: true });
