+++ lib/plugins/aws/deploy/lib/createStack.js
@@ -17,14 +17,7 @@ module.exports = {
       templateBody: this.serverless.service.provider.coreCloudFormationTemplate,
     });
 
-    const executeChangeSetParams = {
-      StackName: stackName,
-      ChangeSetName: changeSetName,
-    };
-
-    if (this.serverless.service.provider.disableRollback) {
-      executeChangeSetParams.DisableRollback = this.serverless.service.provider.disableRollback;
-    }
+    const executeChangeSetParams = this.getExecuteChangeSetParams();
 
     // Create new change set
     log.info('Creating new change set');
