+++ lib/plugins/aws/deploy/lib/ensure-valid-bucket-exists.js
@@ -101,14 +101,7 @@ module.exports = {
       templateBody,
     });
 
-    const executeChangeSetParams = {
-      StackName: stackName,
-      ChangeSetName: changeSetName,
-    };
-
-    if (this.serverless.service.provider.disableRollback) {
-      executeChangeSetParams.DisableRollback = this.serverless.service.provider.disableRollback;
-    }
+    const executeChangeSetParams = this.getExecuteChangeSetParams();
 
     // Ensure that previous change set has been removed
     await this.provider.request('CloudFormation', 'deleteChangeSet', {
