+++ lib/plugins/aws/lib/get-create-change-set-params.js
@@ -60,22 +60,19 @@ module.exports = {
       createChangeSetParams.Parameters = this.serverless.service.provider.stackParameters;
     }
 
-    // Settings below are only applied for `update` type
-    if (changeSetType === 'UPDATE') {
-      // Policy must have at least one statement, otherwise no updates would be possible at all
-      if (
-        this.serverless.service.provider.stackPolicy &&
-        Object.keys(this.serverless.service.provider.stackPolicy).length
-      ) {
-        createChangeSetParams.StackPolicyBody = JSON.stringify({
-          Statement: this.serverless.service.provider.stackPolicy,
-        });
-      }
+    // Policy must have at least one statement, otherwise no updates would be possible at all
+    if (
+      this.serverless.service.provider.stackPolicy &&
+      Object.keys(this.serverless.service.provider.stackPolicy).length
+    ) {
+      createChangeSetParams.StackPolicyBody = JSON.stringify({
+        Statement: this.serverless.service.provider.stackPolicy,
+      });
+    }
 
-      if (this.serverless.service.provider.rollbackConfiguration) {
-        createChangeSetParams.RollbackConfiguration =
-          this.serverless.service.provider.rollbackConfiguration;
-      }
+    if (this.serverless.service.provider.rollbackConfiguration) {
+      createChangeSetParams.RollbackConfiguration =
+        this.serverless.service.provider.rollbackConfiguration;
     }
 
     return createChangeSetParams;
