+++ lib/plugins/aws/lib/updateStack.js
@@ -23,14 +23,7 @@ module.exports = {
       templateUrl,
     });
 
-    const executeChangeSetParams = {
-      StackName: stackName,
-      ChangeSetName: changeSetName,
-    };
-
-    if (this.serverless.service.provider.disableRollback) {
-      executeChangeSetParams.DisableRollback = this.serverless.service.provider.disableRollback;
-    }
+    const executeChangeSetParams = this.getExecuteChangeSetParams();
 
     // Create new change set
     this.provider.didCreateService = true;
@@ -76,14 +69,7 @@ module.exports = {
       templateUrl,
     });
 
-    const executeChangeSetParams = {
-      StackName: stackName,
-      ChangeSetName: changeSetName,
-    };
-
-    if (this.serverless.service.provider.disableRollback) {
-      executeChangeSetParams.DisableRollback = this.serverless.service.provider.disableRollback;
-    }
+    const executeChangeSetParams = this.getExecuteChangeSetParams();
 
     // Ensure that previous change set has been removed
     await this.provider.request('CloudFormation', 'deleteChangeSet', {
