+++ lib/plugins/aws/rollback.js
@@ -5,6 +5,7 @@ const updateStack = require('./lib/updateStack');
 const monitorStack = require('./lib/monitorStack');
 const waitForChangeSetCreation = require('./lib/wait-for-change-set-creation');
 const getCreateChangeSetParams = require('./lib/get-create-change-set-params');
+const getExecuteChangeSetParams = require('./lib/get-execute-change-set-params');
 const findAndGroupDeployments = require('./utils/findAndGroupDeployments');
 const ServerlessError = require('../../serverless-error');
 const { style, log, progress, legacy } = require('@serverless/utils/log');
@@ -24,7 +25,8 @@ class AwsRollback {
       updateStack,
       monitorStack,
       waitForChangeSetCreation,
-      getCreateChangeSetParams
+      getCreateChangeSetParams,
+      getExecuteChangeSetParams
     );
 
     this.hooks = {
